replace `ring::trb` with previous `ring::block`
