+++ src/ring/mod.rs
@@ -1,6 +1,3 @@
 //! TRB Ring.
 
 pub mod trb;
-
-
-pub mod block;
