+++ src/ring/trb/transfer.rs
@@ -1,313 +1,125 @@
 //! Transfer TRBs.
 
-use super::{Link, Type};
 use bit_field::BitField;
 // use core::convert::TryInto;
 use num_derive::FromPrimitive;
 use num_traits::FromPrimitive;
 
-allowed! {
-    /// TRBs which are allowed to be pushed to the Transfer Ring.
-    enum {
-        /// Normal TRB.
-        Normal,
-        /// Setup Stage TRB.
-        SetupStage,
-        /// Data Stage TRB.
-        DataStage,
-        /// Status Stage TRB.
-        StatusStage,
-        /// Isoch TRB.
-        Isoch,
-        /// Link TRB.
-        Link,
-        /// Event Data TRB.
-        EventData,
-        /// No Op TRB.
-        Noop
-    }
-}
-impl Allowed {
-    /// Sets the value of the Interrupt On Completion field.
-    pub fn set_interrupt_on_completion(&mut self) {
-        macro_rules! arm{
-            ($($variant:ident),*)=>{
-                match self {
-                    $(Self::$variant(ref mut x)=>{
-                        x.set_interrupt_on_completion();
-                    },)*
-                }
-            };
-        }
-
-        arm!(
-            Normal,
-            SetupStage,
-            DataStage,
-            StatusStage,
-            Isoch,
-            Link,
-            EventData,
-            Noop
+macro_rules! impl_data_buffer_pointer {
+    () => {
+        rw_double_field!(
+            pub, self,
+            self.0.0; [0, 1],
+            data_buffer_pointer,
+            "Data Buffer Pointer",
+            32, u64
         );
     }
+}
 
-    /// Clears the Interrupt On Completion bit.
-    pub fn clear_interrupt_on_completion(&mut self) {
-        macro_rules! arm{
-            ($($variant:ident),*)=>{
-                match self {
-                    $(Self::$variant(ref mut x)=>{
-                        x.clear_interrupt_on_completion();
-                    },)*
-                }
-            };
-        }
+// 17-bit trb transfer length only.
+macro_rules! impl_trb_transfer_length {
+    () => {
+        rw_field!(pub, self, self.0.0[2]; 0..=16, trb_transfer_length, "TRB Transfer Length", u32);
+    };
+    (prv) => {
+        rw_field!(pub(self), self, self.0.0[2]; 0..=16, trb_transfer_length, "TRB Transfer Length", u32);
+    };
+}
 
-        arm!(
-            Normal,
-            SetupStage,
-            DataStage,
-            StatusStage,
-            Isoch,
-            Link,
-            EventData,
-            Noop
+macro_rules! impl_td_size {
+    () => {
+        rw_field!(pub, self, self.0.0[2]; 17..=21, td_size, "TD Size", u8);
+    };
+}
+
+macro_rules! impl_interrupter_target {
+    () => {
+        rw_field!(pub, self, self.0.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
+    };
+}
+
+macro_rules! impl_ent {
+    () => {
+        rw_bit!(pub, self, self.0.0[3]; 1, evaluate_next_trb, "Evaluate Next TRB");
+    };
+}
+macro_rules! impl_isp {
+    () => {
+        rw_bit!(
+            pub, self,
+            self.0.0[3]; 2,
+            interrupt_on_short_packet,
+            "Interrupt on Short Packet"
         );
-    }
-
-    /// Returns the value of the Interrupt On Completion field.
-    #[must_use]
-    pub fn interrupt_on_completion(&self) -> bool {
-        macro_rules! arm{
-            ($($variant:ident),*)=>{
-                match self {
-                    $(Self::$variant(x)=>{
-                        x.interrupt_on_completion()
-                    },)*
-                }
-            };
-        }
-
-        arm!(
-            Normal,
-            SetupStage,
-            DataStage,
-            StatusStage,
-            Isoch,
-            Link,
-            EventData,
-            Noop
-        )
-    }
+    };
 }
-impl TryFrom<[u32; 4]> for Allowed {
-    type Error = [u32; 4];
-
-    fn try_from(raw: [u32; 4]) -> Result<Self, Self::Error> {
-        try_from!(
-            raw =>
-            Normal,
-            SetupStage,
-            DataStage,
-            StatusStage,
-            Isoch,
-            Link,
-            EventData,
-            Noop(Transfer),
-        );
-        Err(raw)
+macro_rules! impl_ns {
+    () => {
+        rw_bit!(pub, self, self.0.0[3]; 3, no_snoop, "No Snoop");
+    };
+}
+macro_rules! impl_ch {
+    () => {
+        rw_bit!(pub, self, self.0.0[3]; 4, chain_bit, "Chain");
+    };
+}
+macro_rules! impl_idt {
+    () => {
+        rw_bit!(pub, self, self.0.0[3]; 6, immediate_data, "Immediate Data");
+    };
+}
+macro_rules! implprv_idt {
+    () => {
+        rw1s_bit!(pub(self), self, self.0.0[3]; 6, immediate_data, "Immediate Data");
+    };
+}
+macro_rules! impl_bei {
+    () => {
+        rw_bit!(pub, self, self.0.0[3]; 9, block_event_interrupt, "Block Event Interrupt");
+    };
+}
+macro_rules! impl_dir {
+    () => {
+        rw_bit!(pub, self, self.0.0[3]; 16, direction, "Direction (is-in)");
     }
 }
 
-macro_rules! interrupt_on_completion {
-    ($name:ident) => {
-        impl $name {
-            rw_bit!(pub, self, self.0[3]; 5, interrupt_on_completion, "Interrupt On Completion");
-        }
-    };
-}
-macro_rules! transfer_trb {
-    ($name:ident,$full:expr,$type:expr) => {
-        add_trb!($name, $full, $type);
-        interrupt_on_completion!($name);
-    };
-}
-macro_rules! transfer_trb_with_default {
-    ($name:ident,$full:expr,$type:expr) => {
-        add_trb_with_default!($name, $full, $type);
-        interrupt_on_completion!($name);
-    };
-}
-macro_rules! impl_debug_for_transfer_trb{
-    ($name:ident {})=>{
-        impl_debug_for_trb!($name{
-            interrupt_on_completion
-        });
-    };
-    ($name:ident {
-        $($method:ident),*$(,)?
-    })=>{
-        impl_debug_for_trb!($name{
-            interrupt_on_completion,
-            $($method),*
-        });
-    }
-}
-
-transfer_trb_with_default!(Normal, "Normal TRB", Type::Normal);
-reserved!(Normal(Type::Normal) {
-    [3]7..=8;
-    [3]16..=31;
+allowed_trb!("Transfer TRB", {
+    /// Normal TRB
+    Normal = 1,
+    /// Setup Stage TRB
+    "no-new" SetupStage = 2,
+    /// Data Stage TRB
+    DataStage = 3,
+    /// Status Stage TRB
+    StatusStage = 4,
+    /// Isoch TRB
+    Isoch = 5,
+    /// Link TRB
+    Link = 6,
+    /// Event Data TRB
+    EventData = 7,
+    /// No Op Transfer TRB
+    NoOp = 8,
 });
+
 impl Normal {
-    rw_double_field!(
-        pub, self,
-        self.0; [0, 1],
-        data_buffer_pointer,
-        "Data Buffer Pointer",
-        32, u64
-    );
-    rw_field!(pub, self, self.0[2]; 0..=16, trb_transfer_length, "TRB Transfer Length", u32);
-    rw_field!(pub, self, self.0[2]; 17..=21, td_size, "TD Size", u8);
-    rw_field!(pub, self, self.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
-    rw_bit!(pub, self, self.0[3]; 1, evaluate_next_trb, "Evaluate Next TRB");
-    rw_bit!(
-        pub, self,
-        self.0[3]; 2,
-        interrupt_on_short_packet,
-        "Interrupt-on Short Packet"
-    );
-    rw_bit!(pub, self, self.0[3]; 3, no_snoop, "No Snoop");
-    rw_bit!(pub, self, self.0[3]; 4, chain_bit, "Chain bit");
-    rw_bit!(pub, self, self.0[3]; 6, immediate_data, "Immediate Data");
-    rw_bit!(pub, self, self.0[3]; 9, block_event_interrupt, "Block Event Interrupt");
+    impl_data_buffer_pointer!();
+    
+    impl_trb_transfer_length!();
+    impl_td_size!();
+    impl_interrupter_target!();
+
+    impl_ent!();
+    impl_isp!();
+    impl_ns!();
+    impl_ch!();
+    impl_ioc!();
+    impl_idt!();
+    impl_bei!();
 }
-impl_debug_for_transfer_trb! {
-    Normal {
-        data_buffer_pointer,
-        trb_transfer_length,
-        td_size,
-        interrupter_target,
-        cycle_bit,
-        evaluate_next_trb,
-        interrupt_on_short_packet,
-        no_snoop,
-        chain_bit,
-        interrupt_on_completion,
-        immediate_data,
-        block_event_interrupt,
-    }
-}
-
-transfer_trb!(SetupStage, "Setup Stage TRB", Type::SetupStage);
-reserved!(SetupStage(Type::SetupStage) {
-    [2]17..=21;
-    [3]1..=4;
-    [3]7..=9;
-    [3]18..=31;
-});
-impl SetupStage {
-    /// Creates a new Setup Stage TRB.
-    ///
-    /// This method sets the value of the TRB Type, TRB Transfer Length, and the Immediate Data field properly. All the
-    /// other fields are set to 0.
-    #[must_use]
-    pub fn new() -> Self {
-        *Self([0; 4])
-            .set_trb_type()
-            .set_idt()
-            .set_trb_transfer_length()
-    }
-
-    rw_field!(pub, self, self.0[0]; 0..=7, request_type, "bmRequestType", u8);
-    rw_field!(pub, self, self.0[0]; 8..=15, request, "bRequest", u8);
-    rw_field!(pub, self, self.0[0]; 16..=31, value, "wValue", u16);
-    rw_field!(pub, self, self.0[1]; 0..=15, index, "wIndex", u16);
-    rw_field!(pub, self, self.0[1]; 16..=31, length, "wLength", u16);
-    rw_field!(pub, self, self.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
-
-    /// Sets the value of the Transfer Type field.
-    pub fn set_transfer_type(&mut self, t: TransferType) -> &mut Self {
-        self.0[3].set_bits(16..=17, t as _);
-        self
-    }
-
-    /// Returns the value of the Transfer Type field.
-    ///
-    /// # Panics
-    ///
-    /// This method panics if the Transfer Type field contains 1 which is reserved.
-    #[must_use]
-    pub fn transfer_type(&self) -> TransferType {
-        FromPrimitive::from_u32(self.0[3].get_bits(16..=17)).expect("Transfer Type 1 is reserved.")
-    }
-
-    fn set_idt(&mut self) -> &mut Self {
-        self.0[3].set_bit(6, true);
-        self
-    }
-
-    fn set_trb_transfer_length(&mut self) -> &mut Self {
-        self.0[2].set_bits(0..=16, 8);
-        self
-    }
-}
-impl Default for SetupStage {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-impl_debug_for_transfer_trb!(SetupStage {
-    request_type,
-    request,
-    value,
-    index,
-    length,
-    interrupt_on_completion,
-    transfer_type,
-});
-
-transfer_trb_with_default!(DataStage, "Data Stage TRB", Type::DataStage);
-reserved!(DataStage(Type::DataStage) {
-    [3]7..=9;
-    [3]17..=31;
-});
-impl DataStage {
-    rw_double_field!(
-        pub, self,
-        self.0; [0, 1],
-        data_buffer_pointer,
-        "Data Buffer Pointer",
-        32, u64
-    );
-    rw_field!(pub, self, self.0[2]; 0..=16, trb_transfer_length, "TRB Transfer Length", u32);
-    rw_field!(pub, self, self.0[2]; 17..=21, td_size, "TD Size", u8);
-    rw_field!(pub, self, self.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
-    rw_bit!(pub, self, self.0[3]; 1, evaluate_next_trb, "Evaluate Next TRB");
-    rw_bit!(
-        pub, self,
-        self.0[3]; 2,
-        interrupt_on_short_packet,
-        "Interrupt-on Short Packet"
-    );
-    rw_bit!(pub, self, self.0[3]; 3, no_snoop, "No Snoop");
-    rw_bit!(pub, self, self.0[3]; 4, chain_bit, "Chain bit");
-    rw_bit!(pub, self, self.0[3]; 6, immediate_data, "Immediate Data");
-
-    /// Sets the value of the Direction field.
-    pub fn set_direction(&mut self, d: Direction) -> &mut Self {
-        self.0[3].set_bit(16, d.into());
-        self
-    }
-
-    /// Returns the value of the Direction field.
-    #[must_use]
-    pub fn direction(&self) -> Direction {
-        self.0[3].get_bit(16).into()
-    }
-}
-impl_debug_for_transfer_trb!(DataStage {
+impl_debug_from_methods!(Normal {
     data_buffer_pointer,
     trb_transfer_length,
     td_size,
@@ -318,70 +130,155 @@ impl_debug_for_transfer_trb!(DataStage {
     chain_bit,
     interrupt_on_completion,
     immediate_data,
-    direction
+    block_event_interrupt,
+});
+rsvdz_checking_try_from!(Normal {
+    [3];7..=8,
+    [3];16..=31,
 });
 
-transfer_trb_with_default!(StatusStage, "Status Stage TRB", Type::StatusStage);
-reserved!(StatusStage(Type::StatusStage) {
-    [0]0..=31;
-    [1]0..=31;
-    [2]0..=21;
-    [3]2..=3;
-    [3]6..=9;
-    [3]17..=31;
-});
-impl StatusStage {
-    rw_field!(pub, self, self.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
-    rw_bit!(pub, self, self.0[3]; 1, evaluate_next_trb, "Evaluate Next TRB");
-    rw_bit!(pub, self, self.0[3]; 4, chain_bit, "Chain bit");
-    rw_bit!(pub, self, self.0[3]; 16, direction, "Direction");
-}
-impl_debug_for_transfer_trb! {
-    StatusStage {
-        interrupter_target,
-        evaluate_next_trb,
-        chain_bit,
-        interrupt_on_completion,
-        direction,
+impl SetupStage {
+    /// Creates a new Setup Stage TRB.
+    ///
+    /// This method sets the value of the TRB Type, TRB Transfer Length, and the Immediate Data field properly. All the other fields are set to 0.
+    #[must_use]
+    pub fn new() -> Self {
+        *Self(TRB::new(AllowedType::SetupStage))
+            .set_immediate_data()
+            .set_trb_transfer_length(8)
+    }
+
+    rw_field!(pub, self, self.0.0[0]; 0..=7, request_type, "bmRequestType", u8);
+    rw_field!(pub, self, self.0.0[0]; 8..=15, request, "bRequest", u8);
+    rw_field!(pub, self, self.0.0[0]; 16..=31, value, "wValue", u16);
+    rw_field!(pub, self, self.0.0[1]; 0..=15, index, "wIndex", u16);
+    rw_field!(pub, self, self.0.0[1]; 16..=31, length, "wLength", u16);
+
+    impl_trb_transfer_length!(prv);
+    impl_interrupter_target!();
+
+    impl_ioc!();
+    implprv_idt!();
+
+    /// Sets the value of the Transfer Type field.
+    pub fn set_transfer_type(&mut self, t: TransferType) -> &mut Self {
+        self.0.0[3].set_bits(16..=17, t as _);
+        self
+    }
+
+    /// Returns the value of the Transfer Type field.
+    ///
+    /// # Panics
+    ///
+    /// This method panics if the Transfer Type field contains 1 which is reserved.
+    #[must_use]
+    pub fn transfer_type(&self) -> TransferType {
+        FromPrimitive::from_u32(self.0.0[3].get_bits(16..=17))
+            .expect("Transfer Type 1 is reserved.")
     }
 }
+impl_debug_from_methods!(SetupStage {
+    request_type,
+    request,
+    value,
+    index,
+    length,
+    trb_transfer_length,
+    interrupter_target,
+    interrupt_on_completion,
+    immediate_data, // always true
+    transfer_type,
+});
+rsvdz_checking_try_from!(SetupStage { // this won't check IDT and transfer length field.
+    [2];17..=21,
+    [3];1..=4,
+    [3];7..=9,
+    [3];18..=31,
+});
 
-transfer_trb_with_default!(Isoch, "Isoch TRB", Type::Isoch);
-reserved!(Isoch(Type::Isoch) {});
+impl DataStage {
+    impl_data_buffer_pointer!();
+
+    impl_trb_transfer_length!();
+    impl_td_size!();
+    impl_interrupter_target!();
+
+    impl_ent!();
+    impl_isp!();
+    impl_ns!();
+    impl_ch!();
+    impl_ioc!();
+    impl_idt!();
+    impl_dir!();
+}
+impl_debug_from_methods!(DataStage {
+    data_buffer_pointer,
+    trb_transfer_length,
+    td_size,
+    interrupter_target,
+    evaluate_next_trb,
+    interrupt_on_short_packet,
+    no_snoop,
+    chain_bit,
+    interrupt_on_completion,
+    immediate_data,
+    direction,
+});
+rsvdz_checking_try_from!(DataStage {
+    [3];7..=9,
+    [3];17..=31,
+});
+
+impl StatusStage {
+    impl_interrupter_target!();
+
+    impl_ent!();
+    impl_ch!();
+    impl_ioc!();
+    impl_dir!();
+}
+impl_debug_from_methods!(StatusStage {
+    interrupter_target,
+    evaluate_next_trb,
+    chain_bit,
+    interrupt_on_completion,
+    direction,
+});
+rsvdz_checking_try_from!(StatusStage {
+    [0];0..=31,
+    [1];0..=31,
+    [2];0..=21,
+    [3];2..=3,
+    [3];6..=9,
+    [3];17..=31,
+});
+
 impl Isoch {
-    rw_double_field!(
-        pub, self,
-        self.0; [0, 1],
-        data_buffer_pointer,
-        "Data Buffer Pointer",
-        32, u64
-    );
-    rw_field!(pub, self, self.0[2]; 0..=16, trb_transfer_length, "TRB Transfer Length", u32);
-    rw_field!(pub, self, self.0[2]; 17..=21, td_size_or_tbc, "TD Size/TBC", u8);
-    rw_field!(pub, self, self.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
-    rw_bit!(pub, self, self.0[3]; 1, evaluate_next_trb, "Evaluate Next TRB");
-    rw_bit!(
-        pub, self,
-        self.0[3]; 2, 
-        interrupt_on_short_packet,
-        "Interrupt on Short Packet"
-    );
-    rw_bit!(pub, self, self.0[3]; 3, no_snoop, "No Snoop");
-    rw_bit!(pub, self, self.0[3]; 4, chain_bit, "Chain bit");
-    rw_bit!(pub, self, self.0[3]; 6, immediate_data, "Immediate Data");
-    rw_field!(pub, self, self.0[3]; 7..=8, transfer_burst_count, "Transfer Burst Count", u8);
-    rw_bit!(pub, self, self.0[3]; 9, block_event_interrupt, "Block Event Interrupt");
+    impl_data_buffer_pointer!();
+
+    impl_trb_transfer_length!();
+    rw_field!(pub, self, self.0.0[2]; 17..=21, td_size_or_tbc, "TD Size/TBC", u8);
+    impl_interrupter_target!();
+
+    impl_ent!();
+    impl_isp!();
+    impl_ns!();
+    impl_ch!();
+    impl_ioc!();
+    impl_idt!();
+    rw_field!(pub, self, self.0.0[3]; 7..=8, tbc_or_sts, "TBC/TRBSts", u8);
+    impl_bei!();
     rw_field!(
         pub, self,
-        self.0[3]; 16..=19,
+        self.0.0[3]; 16..=19,
         transfer_last_burst_packet_count,
         "Transfer Last Burst Packet Count",
         u8
     );
-    rw_field!(pub, self, self.0[3]; 20..=30, frame_id, "Frame ID", u16);
-    rw_bit!(pub, self, self.0[3]; 31, start_isoch_asap, "Start Isoch ASAP");
+    rw_field!(pub, self, self.0.0[3]; 20..=30, frame_id, "Frame ID", u16);
+    rw_bit!(pub, self, self.0.0[3]; 31, start_isoch_asap, "Start Isoch ASAP");
 }
-impl_debug_for_transfer_trb!(Isoch {
+impl_debug_from_methods!(Isoch {
     data_buffer_pointer,
     trb_transfer_length,
     td_size_or_tbc,
@@ -390,87 +287,89 @@ impl_debug_for_transfer_trb!(Isoch {
     interrupt_on_short_packet,
     no_snoop,
     chain_bit,
+    interrupt_on_completion,
     immediate_data,
-    transfer_burst_count,
+    tbc_or_sts,
     block_event_interrupt,
     transfer_last_burst_packet_count,
     frame_id,
-    start_isoch_asap
+    start_isoch_asap,
+});
+rsvdz_checking_try_from!(Isoch {});
+
+impl Link {
+    impl_ring_segment_pointer!();
+
+    impl_interrupter_target!();
+    
+    impl_tc!();
+    impl_ch!();
+    impl_ioc!();
+}
+rsvdz_checking_try_from!(Link {
+    [0];0..=3,
+    [2];0..=21,
+    [3];2..=3,
+    [3];6..=9,
+    [3];16..=31,
+});
+impl_debug_from_methods!(Link {
+    ring_segment_pointer,
+    interrupter_target,
+    toggle_cycle,
+    chain_bit,
+    interrupt_on_completion,
 });
 
-transfer_trb_with_default!(EventData, "Event Data TRB", Type::EventData);
-reserved!(EventData(Type::EventData) {
-    [2]0..=21;
-    [3]2..=3;
-    [3]6..=8;
-    [3]16..=31;
-});
 impl EventData {
     rw_double_field!(
         pub, self,
-        self.0; [0, 1],
+        self.0.0; [0, 1],
         event_data,
         "Event Data",
         32, u64
     );
-    rw_field!(pub, self, self.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
-    rw_bit!(pub, self, self.0[3]; 1, evaluate_next_trb, "Evaluate Next TRB");
-    rw_bit!(pub, self, self.0[3]; 4, chain_bit, "Chain bit");
-    rw_bit!(pub, self, self.0[3]; 9, block_event_interrupt, "Block Event Interrupt");
+    impl_interrupter_target!();
+    impl_ent!();
+    impl_ch!();
+    impl_ioc!();
+    impl_bei!();
 }
-impl_debug_for_transfer_trb!(EventData {
+impl_debug_from_methods!(EventData {
     event_data,
     interrupter_target,
     evaluate_next_trb,
     chain_bit,
-    block_event_interrupt
+    interrupt_on_completion,
+    block_event_interrupt,
+});
+rsvdz_checking_try_from!(EventData {
+    [2];0..=21,
+    [3];2..=3,
+    [3];6..=8,
+    [3];16..=31,
 });
 
-transfer_trb_with_default!(Noop, "No Op TRB", Type::NoopTransfer);
-reserved!(Noop(Type::NoopTransfer) {
-    [0]0..=31;
-    [1]0..=31;
-    [2]0..=21;
-    [3]2..=3;
-    [3]6..=9;
-    [3]16..=31;
-});
-impl Noop {
-    rw_field!(pub, self, self.0[2]; 22..=31, interrupter_target, "Interrupter Target", u16);
-    rw_bit!(pub, self, self.0[3]; 1, evaluate_next_trb, "Evaluate Next TRB");
-    rw_bit!(pub, self, self.0[3]; 4, chain_bit, "Chain bit");
+impl NoOp {
+    impl_interrupter_target!();
+    impl_ent!();
+    impl_ch!();
+    impl_ioc!();
 }
-impl_debug_for_transfer_trb!(Noop {
+impl_debug_from_methods!(NoOp {
     interrupter_target,
     evaluate_next_trb,
-    chain_bit
+    chain_bit,
+    interrupt_on_completion,
 });
-
-/// The direction of the data transfer.
-#[derive(Copy, Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Hash, FromPrimitive)]
-pub enum Direction {
-    /// Out (Write Data)
-    Out = 0,
-    /// In (Read Data)
-    In = 1,
-}
-impl From<Direction> for bool {
-    fn from(d: Direction) -> Self {
-        match d {
-            Direction::Out => false,
-            Direction::In => true,
-        }
-    }
-}
-impl From<bool> for Direction {
-    fn from(b: bool) -> Self {
-        if b {
-            Direction::In
-        } else {
-            Direction::Out
-        }
-    }
-}
+rsvdz_checking_try_from!(NoOp {
+    [0];0..=31,
+    [1];0..=31,
+    [2];0..=21,
+    [3];2..=3,
+    [3];6..=9,
+    [3];16..=31,
+});
 
 /// Transfer Type.
 #[derive(Copy, Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Hash, FromPrimitive)]
@@ -482,27 +381,4 @@ pub enum TransferType {
     Out = 2,
     /// In Data Stage.
     In = 3,
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn normal_data_buffer_pointer() {
-        let mut normal = Normal::new();
-        let pointer = 0x12345678_9abcdef0;
-        normal.set_data_buffer_pointer(pointer);
-        let pointer_read = normal.data_buffer_pointer();
-        assert_eq!(pointer, pointer_read);
-    }
-
-    #[test]
-    fn isoch_data_buffer_pointer() {
-        let mut isoch = Isoch::new();
-        let pointer = 0xabcd1234_567890ef;
-        isoch.set_data_buffer_pointer(pointer);
-        let pointer_read = isoch.data_buffer_pointer();
-        assert_eq!(pointer, pointer_read);
-    }
 }
